+++ domain/model/user/customer.go
@@ -1,7 +1,7 @@
 package model
 
 type Customer struct {
-	UserID          int    `form:"user_id" json:"user_id"`
+	UserID          int64  `form:"user_id" json:"user_id"`
 	User            User   `form:"user" json:"user"`
 	CustomerAddress string `form:"cust_address" json:"cust_address"`
 	Balance         int    `form:"balance" json:"balance"`
